adding shopping cart functionality

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
 	async createItem(parent, args, ctx, info) {
@@ -47,10 +48,16 @@ const Mutations = {
 
 	async deleteItem(parent, args, ctx, info) {
 		const where = { id: args.id };
-		const item = await ctx.db.query.item({ where }, `{id title}`);
-		// find the item
-		// check if they own the item/have permissions
-		//TODO
+		// 1. find the item
+		const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
+		// 2. check if they own the item, or have permissions
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some(permission =>
+			["ADMIN", "ITEMDELETE"].includes(permission)
+		);
+		if (!ownsItem && !hasPermissions) {
+			throw new Error("You don't have permission to do that!");
+		}
 		// delete item
 		return ctx.db.mutation.deleteItem({ where }, info);
 	},
@@ -181,6 +188,38 @@ const Mutations = {
 		});
 		// 8. return new user
 		return updatedUser;
+	},
+
+	async updatePermissions(parent, args, ctx, info) {
+		// 1. Check if they are logged in
+		if (!ctx.request.userId) {
+			throw new Error("You must be logged in!");
+		}
+		// 2. Query the current user
+		const currentUser = await ctx.db.query.user(
+			{
+				where: {
+					id: ctx.request.userId
+				}
+			},
+			info
+		);
+		// 3. Check if they have permissions to do this
+		hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+		// 4. Update the permissions
+		return ctx.db.mutation.updateUser(
+			{
+				data: {
+					permissions: {
+						set: args.permissions
+					}
+				},
+				where: {
+					id: args.userId
+				}
+			},
+			info
+		);
 	}
 };
